requirements.txt
dash==2.3.1
dash-bootstrap-components==1.1.0
numpy==1.21.6
pandas==1.3.5
Pillow==9.1.0
plotly==5.7.0
scipy==1.7.3

